% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{.entropy_ghype}
\alias{.entropy_ghype}
\title{Entropy of multinomial approximation.}
\usage{
.entropy_ghype(
  ensemble = NULL,
  xi = NULL,
  omega = NULL,
  directed = NULL,
  selfloops = NULL,
  m = NULL
)
}
\description{
Entropy of multinomial approximation.
}
