% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{.max_entropy}
\alias{.max_entropy}
\title{Maximum entropy of multinomial approximation.}
\usage{
.max_entropy(
  ensemble = NULL,
  m = NULL,
  n = NULL,
  directed = NULL,
  selfloops = NULL
)
}
\description{
Maximum entropy of multinomial approximation.
}
