% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{entropy_ghype}
\alias{entropy_ghype}
\title{Entropy of multinomial approximation.}
\usage{
entropy_ghype(
  ensemble = NULL,
  xi = NULL,
  omega = NULL,
  directed = NULL,
  selfloops = NULL,
  m = NULL
)
}
\arguments{
\item{ensemble}{ghype ensemble whose entropy to compute.}

\item{xi}{A combinatorial matrix Xi to use instead of the ensemble's.}

\item{omega}{A propensity matrix Omega to use instead of the ensemble's.}

\item{directed}{Whether ensemble contains directed networks.}

\item{selfloops}{Whether ensemble contains networks with selfloops.}

\item{m}{Number of edges in a network in ensemble.}
}
\value{
Computed entropy.
}
\description{
Entropy of multinomial approximation.
}
