% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{potentiality.ghype}
\alias{potentiality.ghype}
\title{Computes the potentiality for a given ghype ensemble.}
\usage{
\method{potentiality}{ghype}(x, ...)
}
\arguments{
\item{x}{ghype ensemble whose potentiality to compute.}

\item{...}{Further arguments are ignored.}
}
\value{
Computed potentiality in the interval \verb{[0, 1]}.
}
\description{
Computes the potentiality for a given ghype ensemble.
}
