% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{potentiality.igraph}
\alias{potentiality.igraph}
\title{Computes the potentiality according to a MLE fit to a given network.}
\usage{
\method{potentiality}{igraph}(
  x,
  directed = igraph::is_directed(x),
  selfloops = any(igraph::which_loop(x)),
  full_model = TRUE,
  ...
)
}
\arguments{
\item{x}{Multi-edge network of which to compute the potentiality, given as an
igraph.}

\item{directed}{Whether x is directed. If omitted, this is detected
from base_network.}

\item{selfloops}{Whether base_network is allowed to have self-loops. If
omitted, this is detected from base_network.}

\item{full_model}{Whether to use a full ghype (TRUE) or bccm (FALSE).
Defaults to TRUE.}

\item{...}{Further arguments are ignored.}
}
\value{
Computed potentiality in the interval \verb{[0, 1]}.
}
\description{
The MLE fit has the property that x is preserved as the expected network over
the ensemble.
}
\details{
For the special cases where x has no nodes or no edges, a potentiality of 0
is returned because there is always only this \emph{one} network which fulfills
these constraints.
}
