% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{Potentiality}
\alias{Potentiality}
\title{Computes the potentiality according to a MLE fit.}
\usage{
Potentiality(
  network,
  directed = igraph::is_directed(network),
  has_selfloops = any(igraph::which_loop(network)),
  full_model = TRUE
)
}
\arguments{
\item{network}{igraph graph of which to compute the potentiality.}

\item{directed}{Whether network is directed. If omitted, this is detected from base_network.}

\item{has_selfloops}{Whether base_network is allowed to have self-loops. If omitted, this is
detected from base_network.}

\item{full_model}{Whether to use a full ghype (TRUE) or bccm (FALSE). Defaults to TRUE.}
}
\value{
The computed potentiality.
}
\description{
The MLE fit has the property that network is preserved as the expected network over the ensemble.
}
\details{
For the special cases where network has no nodes or no edges, a potentiality of 0 is returned
because there is always only this \emph{one} network which fulfils these constraints.
}
