% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{.dbinom_sum_over_probs}
\alias{.dbinom_sum_over_probs}
\title{Compute vector v, where v_i := sum_{p in probs} dbinom(xs_i, size, p)}
\usage{
.dbinom_sum_over_probs(xs, size, probs)
}
\description{
Compute vector v, where v_i := sum_{p in probs} dbinom(xs_i, size, p)
}
