% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{.max_entropy}
\alias{.max_entropy}
\title{Maximum entropy of multinomial approximation.}
\usage{
.max_entropy(ens)
}
\arguments{
\item{ens}{ghype ensemble whose maximum entropy to compute.}
}
\value{
Computed maximum entropy.
}
\description{
Maximum entropy of multinomial approximation.
}
