% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{.relative_entropy}
\alias{.relative_entropy}
\title{Computes the model's entropy normalized by the theoretical maximum.}
\usage{
.relative_entropy(model)
}
\description{
Computes the model's entropy normalized by the theoretical maximum.
}
