% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{entropy_ghype}
\alias{entropy_ghype}
\title{Entropy of multinomial approximation to Wallenius distribution.}
\usage{
entropy_ghype(ens)
}
\arguments{
\item{ens}{ghype ensemble whose entropy to compute.}
}
\value{
Computed entropy.
}
\description{
Entropy of multinomial approximation to Wallenius distribution.
}
