% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentiality.R
\name{potentiality}
\alias{potentiality}
\title{Generic method to compute the potentiality.}
\usage{
potentiality(x, ...)
}
\arguments{
\item{x}{Object whose potentiality to compute. For example, it may be an
igraph or a ghype ensemble.}

\item{...}{Further arguments passed to the methods.}
}
\value{
Computed potentiality in the interval \verb{[0, 1]}.
}
\description{
Generic method to compute the potentiality.
}
